[BEANUTILS-429] - Delete trailing white spaces and white spaces on empty lines from all files
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/beanutils/trunk@1452028 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/beanutils/locale/converters/FloatLocaleConverter.java
@@ -217,7 +217,7 @@ public class FloatLocaleConverter extends DecimalLocaleConverter {
    protected Object parse(Object value, String pattern) throws ParseException {
       final Number parsed = (Number) super.parse(value, pattern);
       double doubleValue = parsed.doubleValue();
-      double posDouble = (doubleValue >= (double)0) ? doubleValue : (doubleValue * (double)-1);
+      double posDouble = (doubleValue >= 0) ? doubleValue : (doubleValue * -1);
       if (posDouble != 0 && (posDouble < Float.MIN_VALUE || posDouble > Float.MAX_VALUE)) {
           throw new ConversionException("Supplied number is not of type Float: "+parsed);
       }
